+++ src/models/__init__.py
@@ -2,6 +2,12 @@
 
 
 from .self_attention import resnet18_self_attention
+from .squeeze_and_excite import (
+    resnet18_se,
+    resnet50_se,
+    resnet18_se_fc512,
+    resnet50_se_fc512,
+)
 from .resnet import (
     resnet18,
     resnet18_se,
@@ -21,18 +27,19 @@ from .tvmodels import mobilenet_v3_small, vgg16
 __model_factory = {
     # image classification models
     "resnet18": resnet18,
-    "resnet18_se": resnet18_se,
-    "resnet18_self_attention": resnet18_self_attention,
     "resnet18_fc512": resnet18_fc512,
-    "resnet18_se_fc512": resnet18_se_fc512,
     "resnet34": resnet34,
     "resnet34_fc512": resnet34_fc512,
     "resnet50": resnet50,
-    "resnet50_se": resnet50_se,
     "resnet50_fc512": resnet50_fc512,
-    "resnet50_se_fc512": resnet50_se_fc512,
     "mobilenet_v3_small": mobilenet_v3_small,
     "vgg16": vgg16,
+    # custom models
+    "resnet18_se": resnet18_se,
+    "resnet18_se_fc512": resnet18_se_fc512,
+    "resnet18_self_attention": resnet18_self_attention,
+    "resnet50_se": resnet50_se,
+    "resnet50_se_fc512": resnet50_se_fc512,
 }
