+++ src/models/resnet.py
@@ -2,19 +2,19 @@
 
 import torch.utils.model_zoo as model_zoo
 from torch import nn
-import torch
 
 __all__ = [
     "resnet18",
-    "resnet18_se",
-    "resnet18_self_attention",
     "resnet18_fc512",
-    "resnet18_se_fc512",
     "resnet34",
     "resnet34_fc512",
     "resnet50",
-    "resnet50_se",
     "resnet50_fc512",
+    # custom implementation
+    "resnet18_se",
+    "resnet18_se_fc512",
+    "resnet18_self_attention",
+    "resnet50_se",
     "resnet50_se_fc512",
 ]
 
@@ -66,39 +66,6 @@ class BasicBlock(nn.Module):
         return out
 
 
-class SEBasicBlock(BasicBlock):
-    def __init__(self, inplanes, planes, stride=1, downsample=None, reduction=16):
-        super(SEBasicBlock, self).__init__(inplanes, planes, stride, downsample)
-        self.se = nn.Sequential(
-            nn.AdaptiveAvgPool2d(1),
-            nn.Conv2d(planes, planes // reduction, kernel_size=1),
-            nn.ReLU(inplace=True),
-            nn.Conv2d(planes // reduction, planes, kernel_size=1),
-            nn.Sigmoid(),
-        )
-
-    def forward(self, x):
-        residual = x
-
-        out = self.conv1(x)
-        out = self.bn1(out)
-        out = self.relu(out)
-
-        out = self.conv2(out)
-        out = self.bn2(out)
-
-        # Apply SE attention
-        out = out * self.se(out)
-
-        if self.downsample is not None:
-            residual = self.downsample(x)
-
-        out += residual
-        out = self.relu(out)
-
-        return out
-
-
 class Bottleneck(nn.Module):
     expansion = 4
 
@@ -139,27 +106,6 @@ class Bottleneck(nn.Module):
         out = self.relu(out)
 
         return out
-
-
-class SEBottleneck(Bottleneck):
-    def __init__(self, inplanes, planes, stride=1, downsample=None, reduction=16):
-        super(SEBottleneck, self).__init__(inplanes, planes, stride, downsample)
-
-        self.se = nn.Sequential(
-            nn.AdaptiveAvgPool2d(1),
-            nn.Conv2d(
-                planes * self.expansion,
-                planes * self.expansion // reduction,
-                kernel_size=1,
-            ),
-            nn.ReLU(inplace=True),
-            nn.Conv2d(
-                planes * self.expansion // reduction,
-                planes * self.expansion,
-                kernel_size=1,
-            ),
-            nn.Sigmoid(),
-        )
 
     def forward(self, x):
         residual = x
@@ -370,43 +316,11 @@ def resnet18(num_classes, loss={"xent"}, pretrained=True, **kwargs):
     return model
 
 
-def resnet18_se(num_classes, loss={"xent"}, pretrained=True, **kwargs):
-    model = ResNet(
-        num_classes=num_classes,
-        loss=loss,
-        block=SEBasicBlock,
-        layers=[2, 2, 2, 2],
-        last_stride=2,
-        fc_dims=None,
-        dropout_p=None,
-        **kwargs,
-    )
-    if pretrained:
-        init_pretrained_weights(model, model_urls["resnet18"])
-    return model
-
-
 def resnet18_fc512(num_classes, loss={"xent"}, pretrained=True, **kwargs):
     model = ResNet(
         num_classes=num_classes,
         loss=loss,
         block=BasicBlock,
-        layers=[2, 2, 2, 2],
-        last_stride=1,
-        fc_dims=[512],
-        dropout_p=None,
-        **kwargs,
-    )
-    if pretrained:
-        init_pretrained_weights(model, model_urls["resnet18"])
-    return model
-
-
-def resnet18_se_fc512(num_classes, loss={"xent"}, pretrained=True, **kwargs):
-    model = ResNet(
-        num_classes=num_classes,
-        loss=loss,
-        block=SEBasicBlock,
         layers=[2, 2, 2, 2],
         last_stride=1,
         fc_dims=[512],
@@ -466,43 +380,11 @@ def resnet50(num_classes, loss={"xent"}, pretrained=True, **kwargs):
     return model
 
 
-def resnet50_se(num_classes, loss={"xent"}, pretrained=True, **kwargs):
-    model = ResNet(
-        num_classes=num_classes,
-        loss=loss,
-        block=SEBottleneck,
-        layers=[3, 4, 6, 3],
-        last_stride=2,
-        fc_dims=None,
-        dropout_p=None,
-        **kwargs,
-    )
-    if pretrained:
-        init_pretrained_weights(model, model_urls["resnet50"])
-    return model
-
-
 def resnet50_fc512(num_classes, loss={"xent"}, pretrained=True, **kwargs):
     model = ResNet(
         num_classes=num_classes,
         loss=loss,
         block=Bottleneck,
-        layers=[3, 4, 6, 3],
-        last_stride=1,
-        fc_dims=[512],
-        dropout_p=None,
-        **kwargs,
-    )
-    if pretrained:
-        init_pretrained_weights(model, model_urls["resnet50"])
-    return model
-
-
-def resnet50_se_fc512(num_classes, loss={"xent"}, pretrained=True, **kwargs):
-    model = ResNet(
-        num_classes=num_classes,
-        loss=loss,
-        block=SEBottleneck,
         layers=[3, 4, 6, 3],
         last_stride=1,
         fc_dims=[512],
